requirements.txt
requests
elevenlabs
groq
Pillow
opencv-python
google-generativeai
pyHS100
mss
pyautogui
keyboard
google-auth-oauthlib
google-api-python-client
google-auth 
termcolor
SpeechRecognition
pydub
pyaudio
vosk
pyqt5
pygame
pynput
flask
